Fixed ClimbCommand
