+++ ELON_HUSK/src/main/java/frc/robot/commands/climb/ClimbDownCommand.java
@@ -1,15 +1,25 @@
 package frc.robot.commands.climb;
 
+import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.subsystems.ClimbSubsystem;
 
-public class ClimbDownCommand extends ClimbCommand {
+public class ClimbDownCommand extends CommandBase {
 
-    public ClimbDownCommand(ClimbSubsystem climb){
-        super(climb);
+	protected final ClimbSubsystem _climb;
+
+	public ClimbDownCommand(ClimbSubsystem climb) {
+		_climb = climb;
+		addRequirements(climb);
+	}
+
+	@Override
+	public void initialize() {
+		_climb.setInputDirection(ClimbSubsystem.InputDirection.DOWN);
+	}
+
+	@Override
+    public void end(boolean interrupted) {
+        _climb.setInputDirection(ClimbSubsystem.InputDirection.NEUTRAL);
     }
 
-    @Override
-    public void initialize() {
-        _climb.setInputDirection(ClimbSubsystem.InputDirection.DOWN);
-    }
 }
