+++ ELON_HUSK/src/main/java/frc/robot/commands/climb/ClimbUpCommand.java
@@ -1,15 +1,25 @@
 package frc.robot.commands.climb;
 
+import edu.wpi.first.wpilibj2.command.CommandBase;
 import frc.robot.subsystems.ClimbSubsystem;
 
-public class ClimbUpCommand extends ClimbCommand {
+public class ClimbUpCommand extends CommandBase {
 
-    public ClimbUpCommand(ClimbSubsystem climb){
-        super(climb);
+	protected final ClimbSubsystem _climb;
+
+	public ClimbUpCommand(ClimbSubsystem climb) {
+		_climb = climb;
+		addRequirements(climb);
+	}
+
+	@Override
+	public void initialize() {
+		_climb.setInputDirection(ClimbSubsystem.InputDirection.UP);
+	}
+
+	@Override
+    public void end(boolean interrupted) {
+        _climb.setInputDirection(ClimbSubsystem.InputDirection.NEUTRAL);
     }
 
-    @Override
-    public void initialize() {
-        _climb.setInputDirection(ClimbSubsystem.InputDirection.UP);
-    }
 }
